feat : 노드맵 api SSR적용
