+++ Mindspace_front_next/app/map/hydratedNode.tsx
@@ -1,0 +1,19 @@
+import { dehydrate } from "@tanstack/react-query";
+import getQueryClient from "@/getQueryClient";
+import { getNodeList } from "@/api/node";
+import { NODE_QUERIES } from "@/constants/queryKeys";
+import MapPage from "./map";
+import HydrateOnClient from "@/hydrateOnClient";
+
+export default async function HydratedNode() {
+  const queryClient = getQueryClient();
+  await queryClient.prefetchQuery([NODE_QUERIES.LIST], getNodeList);
+
+  const dehydratedState = dehydrate(queryClient);
+
+  return (
+    <HydrateOnClient state={dehydratedState}>
+      <MapPage />
+    </HydrateOnClient>
+  );
+}
